# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Scantour API",
        "version": "0.1"
    },
    "paths": {
        "/api/excursions": {
            "get": {
                "tags": [
                    "Excursions"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "404": {
                        "description": "Excursion not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/excursions/{id}": {
            "get": {
                "tags": [
                    "Excursions"
                ],
                "summary": "Get excursion by id",
                "operationId": "6a134d4edaed1bebdfd99ca69923de72",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of excursion",
                        "required": true,
                        "example": "7"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "404": {
                        "description": "Excursion not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/excursions/availability": {
            "get": {
                "tags": [
                    "Excursions"
                ],
                "summary": "Получение доступности и цен по всем экскурсиям. Параметр id соответствует tariffs->id в методе /api/excursions/{id}",
                "operationId": "fc8d28073c22927ffa536599f5197a8e",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "404": {
                        "description": "Excursions Availability not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "description": "ДЛя передачи заказов, необходимо отправить POST запрос к методу api/orders . <br /> POST-запрос должен содержать ассоциативный массив.<br /><br /> <Strong>Пример реализации запроса на PHP:</strong>\n     * <br />\n     *  <pre class='hljs' style='display: block; overflow-x: auto; padding: 0.5em; background: rgb(240, 240, 240); color: rgb(68, 68, 68);'>$data = <span class='hljs-keyword' style='font-weight: 700;'>array</span>();</pre>\n     * ",
                "operationId": "d690d6834318e9b064df441e854de8b9",
                "parameters": [
                    {
                        "name": "login",
                        "in": "query",
                        "description": "Login",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true
                    },
                    {
                        "name": "tarif",
                        "in": "query",
                        "description": "Number of tarif. Ex. 1 - Hotel Standart, 2 - Hotel Comfort etc.",
                        "required": true
                    },
                    {
                        "name": "raceID",
                        "in": "query",
                        "description": "Race ID, Ex. 17255",
                        "required": true
                    },
                    {
                        "name": "comment",
                        "in": "query",
                        "description": "Comment",
                        "required": false
                    },
                    {
                        "name": "sngl",
                        "in": "query",
                        "description": "number of single placements, 0 - if not needed",
                        "required": true
                    },
                    {
                        "name": "dbl",
                        "in": "query",
                        "description": "number of double placements, 0 - if not needed",
                        "required": true
                    },
                    {
                        "name": "trpl",
                        "in": "query",
                        "description": "number of triple placements, 0 - if not needed",
                        "required": true
                    },
                    {
                        "name": "qdrpl",
                        "in": "query",
                        "description": "number of quadriple placements, 0 - if not needed",
                        "required": true
                    },
                    {
                        "name": "arr[i]['passport']",
                        "in": "query",
                        "description": "(i - tourist index) passport number",
                        "required": true
                    },
                    {
                        "name": "arr[i]['mail']",
                        "in": "query",
                        "description": "tourist email",
                        "required": true
                    },
                    {
                        "name": "arr[i]['phone']",
                        "in": "query",
                        "description": "tourist phone number",
                        "required": true
                    },
                    {
                        "name": "arr[i]['name']",
                        "in": "query",
                        "description": "tourist name",
                        "required": true
                    },
                    {
                        "name": "arr[i]['category']",
                        "in": "query",
                        "description": "tourist category",
                        "required": true
                    },
                    {
                        "name": "arr[i]['birth_date']",
                        "in": "query",
                        "description": "tourist birth date",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "404": {
                        "description": "Tours not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tours": {
            "get": {
                "tags": [
                    "Tours"
                ],
                "operationId": "5350dec906fe00f938f344ef49511e39",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "404": {
                        "description": "Tours not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tours/{id}": {
            "get": {
                "tags": [
                    "Tours"
                ],
                "summary": "Get tour by id",
                "operationId": "3ca89226e852b677757991cc742f340d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of tour",
                        "required": true,
                        "example": "7"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "404": {
                        "description": "Tour not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tours/availability": {
            "get": {
                "tags": [
                    "Tours"
                ],
                "summary": "Получение доступности и цен по всем турам. Параметр id соответствует accomodation->id в методе /api/tours/{id}",
                "operationId": "7259e565b67b5476abb2e8aae50e0f43",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "404": {
                        "description": "Tours Availability not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Excursion": {
                "title": "Excursions schema",
                "properties": {
                    "id": {
                        "description": "id of excursion\n  * ",
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "type": "object"
            },
            "ExcursionTariff": {
                "title": "Excursion's tariff schema",
                "properties": {
                    "availability": {
                        "description": "2 - Много мест, 1 - По запросу, 0 - нет мест",
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            },
            "0": []
        }
    }
}